[MRG] Automatic merge, trunk 32389 to 35502
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/experimental@35510 b456876b-0849-0410-b77d-98878d47e9d5

+++ coe_fgal_relook2/tiki-setup_base.php
@@ -1,5 +1,5 @@
 <?php
-// (c) Copyright 2002-2010 by authors of the Tiki Wiki/CMS/Groupware Project
+// (c) Copyright 2002-2011 by authors of the Tiki Wiki CMS Groupware Project
 // 
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
 // Licensed under the GNU LESSER GENERAL PUBLIC LICENSE. See license.txt for details.
@@ -50,7 +50,6 @@ $needed_prefs = array(
 	'tiki_cdn_ssl' => '',
 	'language' => 'en',
 	'lang_use_db' => 'n',
-	'feature_pear_date' => 'y',
 	'lastUpdatePrefs' => - 1,
 	'feature_fullscreen' => 'n',
 	'error_reporting_level' => 0,
@@ -60,7 +59,15 @@ $needed_prefs = array(
 	'memcache_prefix' => 'tiki_',
 	'memcache_compress' => 'y',
 	'memcache_servers' => false,
+	'min_pass_length' => 5,
+	'pass_chr_special' => 'n'
 );
+// check that tiki_preferences is there
+if (empty($tikilib->query("SHOW TABLES LIKE 'tiki_preferences'")->numrows)) {
+	// smarty not initialised at this point to do a polite message, sadly
+	header('location: tiki-install.php');
+	exit;
+}
 $tikilib->get_preferences($needed_prefs, true, true);
 if (!isset($prefs['lastUpdatePrefs']) || $prefs['lastUpdatePrefs'] == - 1) {
 	$tikilib->query('delete from `tiki_preferences` where `name`=?', array('lastUpdatePrefs'));
@@ -456,7 +463,10 @@ if (isset($_SESSION["$user_cookie_site"])) {
 	// }
 	
 }
-$smarty->assign( 'CSRFTicket', isset( $_SESSION['ticket'] ) ? $_SESSION['ticket'] : null);
+
+if (is_object($smarty)) {
+	$smarty->assign( 'CSRFTicket', isset( $_SESSION['ticket'] ) ? $_SESSION['ticket'] : null);
+}
 require_once ('lib/setup/perms.php');
 // --------------------------------------------------------------
 // deal with register_globals
@@ -535,6 +545,26 @@ if ($tiki_p_trust_input != 'y') {
 	}
 	unset($tmp);
 }
+
+if ($prefs['tiki_check_file_content'] == 'y' && count($_FILES)) {
+	if ($finfo = new finfo(FILEINFO_MIME)) {
+
+		foreach ($_FILES as $key => & $upload_file_info) {
+			if (is_array($upload_file_info['tmp_name'])) {
+				foreach ($upload_file_info['tmp_name'] as $k => $tmp_name) {
+					if ($tmp_name) {
+						$upload_file_info['type'][$k] = $finfo->file($tmp_name);
+					}
+				}
+			} elseif ($upload_file_info['tmp_name']) {
+				$upload_file_info['type'] = $finfo->file($upload_file_info['tmp_name']);
+			}
+		}
+	}
+
+	unset($finfo);
+}
+
 // deal with old request globals (e.g. used by Smarty)
 $GLOBALS['HTTP_GET_VARS'] = & $_GET;
 $GLOBALS['HTTP_POST_VARS'] = & $_POST;
@@ -556,4 +586,6 @@ if (!isset($_SERVER['QUERY_STRING'])) $_SERVER['QUERY_STRING'] = '';
 if (!isset($_SERVER['REQUEST_URI']) || empty($_SERVER['REQUEST_URI'])) {
 	$_SERVER['REQUEST_URI'] = $_SERVER['PHP_SELF'] . '?' . $_SERVER['QUERY_STRING'];
 }
-$smarty->assign("tikidomain", $tikidomain);
+if (is_object($smarty)) {
+	$smarty->assign("tikidomain", $tikidomain);
+}
